+++ CHANGELOG.md
@@ -4,6 +4,36 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.1.1 - 2025-07-15
+### Fixed
+- missing gateway setting variable
+
+## 3.1.0 - 2025-05-05
+### Added
+- multisite override support for merchant account ids
+- device data collection for dropin-ui
+
+### Updated
+- DropInUiSDK version 1.44.1
+- ClientSDK version 3.111.0
+
+## 3.0.3 - 2024-05-16
+### Added
+- Craft 5 compatibility
+
+### Updated
+- DropInUiSDK version 1.42.0
+- ClientSDK version 3.102.0
+
+## 3.0.2 - 2024-03-13
+### Fixed
+- Braintree\Customer type
+
+## 3.0.1 - 2024-02-06
+### Changed
+- manual state control option for buttons
+- Released - finally!
+
 ## 3.0.0-beta.11 - 2023-05-18
 ### Changed
 - changed callback param to object with callbacks for "onPaymentMethodSelect" and "onPaymentMethodReady"